layer preview files renamed to comply with new approach
